# src/app/api/openapi.yaml
openapi: 3.0.0
info:
  title: ZenZen CRM API
  version: '1.0.0'
  description: |
    API for the ZenZen CRM system, providing ticket management, messaging, and team collaboration features.
  contact:
    name: ZenZen Support
    url: https://zenzen.support
servers:
  - url: https://api.zenzen.dev
    description: Development server
  - url: https://api.zenzen.com
    description: Production server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Team:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        focus_area:
          type: string
        created_at:
          type: string
          format: date-time
...
paths:
...
  /teams:
    get:
      summary: List teams
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of teams
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Team'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      pages:
                        type: integer
                      current_page:
                        type: integer
                      per_page:
                        type: integer

    post:
      summary: Create a new team
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                focus_area:
                  type: string
      responses:
        '201':
          description: Team created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
  /teams/{team_id}/users:
    parameters:
      - name: team_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Add user to team
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
              properties:
                user_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: User added to team successfully

    delete:
      summary: Remove user from team
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User removed from team successfully
...
security:
  - ApiKeyAuth: []
  - BearerAuth: [] 